Update rate_table for SNOwGLoBES 1.3.2
Due to a bug in earlier SNOwGLoBES versions (https://github.com/SNOwGLoBES/snowglobes/pull/28), snewpy wouldn’t load efficiency files for ar40kt and produce this warning:
> python/snewpy/rate_calculator.py:389: UserWarning: Efficiency not found for detector=ar40kt, channel=nue_Ar40_he. Using 100% efficiency
(Note the `_he` in the channel name!)

With snowglobes_data 1.3.2, snewpy finds the correct efficiency file. This leads to a lower efficiency and we lose some events.
This commit updates the rate_table with the new event counts.